+++ pdoc/__init__.py
@@ -379,7 +379,7 @@ You can find an example in [`examples/library-usage`](https://github.com/mitmpro
 '''
 from __future__ import annotations
 
-__version__ = "10.0.3"  # this is read from setup.py
+__version__ = "10.0.4"  # this is read from setup.py
 
 import traceback
 import warnings